+++ src/Component/OpenApi31/Tests/fixtures/enum-as-objects/expected/Model/ItemStatus.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Jane\Component\OpenApi31\Tests\EnumAsObjects\Model;
+
+enum ItemStatus : string
+{
+    case Available = 'available';
+    case OutOfStock = 'out-of-stock';
+    case Reserved = 'reserved';
+    case NewYork = 'new-york';
+    case Newyork = 'NewYork';
+}