+++ src/Component/OpenApi31/Tests/fixtures/enum-as-objects/expected/Model/Priority.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Jane\Component\OpenApi31\Tests\EnumAsObjects\Model;
+
+enum Priority : int
+{
+    case Value1 = 1;
+    case Value2 = 2;
+    case Value3 = 3;
+}